Tweet delete: check if it's within 3 minutes

+++ lib/controllers/tweet.controller.ts
@@ -1,6 +1,7 @@
 import {Request, Response, NextFunction} from 'express';
 import {Tweet} from "../models/tweet.model";
 import {User} from "../models/user.model";
+import {create} from "domain";
 
 export const postAddTweet = (req: Request, res: Response, next: NextFunction) => {
 
@@ -34,23 +35,27 @@ export const getFeed = (req: Request, res: Response, next: NextFunction) => {
             res.send(tweets);
         })
         .catch(err => console.log(err));
-}
+};
+
 export const deleteTweet = (req: Request, res: Response, next: NextFunction) => {
     const tweetId: number = req.body.tweetId;
     // @ts-ignore
 
     Tweet.findByPk(tweetId).then(token => {
 
-        if (token !== null){
-            Tweet.destroy({where: {id: tweetId}});
-            res.status(204).send('Tweet deleted.');
-        }
-        else {
+        if (token !== null) {
+            // check if it's within 3 minutes
+            const createdAt: number = new Date(token['dataValues'].createdAt).getTime();
+            const diffTimeMs: number = Date.now() - createdAt;
+            if (diffTimeMs > 3 * 60 * 1000) {
+                res.status(400).send("Too late!");
+            } else {
+                Tweet.destroy({where: {id: tweetId}});
+                res.status(204).send('Tweet deleted.');
+            }
+        } else {
             res.status(400).send(`Tweet with id ${tweetId} not found.`);
         }
 
     }).catch(err => console.log(err))
-
-
-    // todo check if it's under 3 minutes
 };
